Renamed a number of criterions (#538)
