+++ ta4j-core/src/main/java/org/ta4j/core/analysis/criteria/LinearTransactionCostCriterion.java
@@ -42,7 +42,7 @@ public class LinearTransactionCostCriterion extends AbstractAnalysisCriterion {
     private double a;
     private double b;
 
-    private TotalProfitCriterion profit;
+    private TotalReturnCriterion profit;
 
     /**
      * Constructor. (a * x)
@@ -68,7 +68,7 @@ public class LinearTransactionCostCriterion extends AbstractAnalysisCriterion {
         this.initialAmount = initialAmount;
         this.a = a;
         this.b = b;
-        profit = new TotalProfitCriterion();
+        profit = new TotalReturnCriterion();
     }
 
     @Override
