+++ ta4j-core/src/main/java/org/ta4j/core/analysis/criteria/ReturnOverMaxDrawdownCriterion.java
@@ -23,32 +23,31 @@
  */
 package org.ta4j.core.analysis.criteria;
 
+import org.ta4j.core.AnalysisCriterion;
 import org.ta4j.core.BarSeries;
-import org.ta4j.core.Order;
 import org.ta4j.core.Trade;
 import org.ta4j.core.TradingRecord;
+import org.ta4j.core.num.NaN;
 import org.ta4j.core.num.Num;
 
 /**
- * Buy and hold criterion.
- *
- * Calculates the return if a buy-and-hold strategy was used, buying on the
- * first bar and selling on the last bar.
- *
- * @see <a href=
- *      "http://en.wikipedia.org/wiki/Buy_and_hold">http://en.wikipedia.org/wiki/Buy_and_hold</a>
+ * Reward risk ratio criterion, defined as the {@link TotalReturnCriterion total
+ * profit} over the {@link MaximumDrawdownCriterion maximum drawdown}.
  */
-public class BuyAndHoldCriterion extends AbstractAnalysisCriterion {
+public class ReturnOverMaxDrawdownCriterion extends AbstractAnalysisCriterion {
+
+    private final AnalysisCriterion totalProfitCriterion = new TotalReturnCriterion();
+    private final AnalysisCriterion maxDrawdownCriterion = new MaximumDrawdownCriterion();
 
     @Override
     public Num calculate(BarSeries series, TradingRecord tradingRecord) {
-        return createBuyAndHoldTrade(series).getGrossReturn(series);
-    }
-
-    @Override
-    public Num calculate(BarSeries series, Trade trade) {
-        return createBuyAndHoldTrade(series, trade.getEntry().getIndex(), trade.getExit().getIndex())
-                .getGrossReturn(series);
+        final Num maxDrawdown = maxDrawdownCriterion.calculate(series, tradingRecord);
+        if (maxDrawdown.isZero()) {
+            return NaN.NaN;
+        } else {
+            final Num totalProfit = totalProfitCriterion.calculate(series, tradingRecord);
+            return totalProfit.dividedBy(maxDrawdown);
+        }
     }
 
     @Override
@@ -56,14 +55,14 @@ public class BuyAndHoldCriterion extends AbstractAnalysisCriterion {
         return criterionValue1.isGreaterThan(criterionValue2);
     }
 
-    private Trade createBuyAndHoldTrade(BarSeries series) {
-        return createBuyAndHoldTrade(series, series.getBeginIndex(), series.getEndIndex());
-    }
-
-    private Trade createBuyAndHoldTrade(BarSeries series, int beginIndex, int endIndex) {
-        Trade trade = new Trade(Order.OrderType.BUY);
-        trade.operate(beginIndex, series.getBar(beginIndex).getClosePrice(), series.numOf(1));
-        trade.operate(endIndex, series.getBar(endIndex).getClosePrice(), series.numOf(1));
-        return trade;
+    @Override
+    public Num calculate(BarSeries series, Trade trade) {
+        final Num maxDrawdown = maxDrawdownCriterion.calculate(series, trade);
+        if (maxDrawdown.isZero()) {
+            return NaN.NaN;
+        } else {
+            final Num totalProfit = totalProfitCriterion.calculate(series, trade);
+            return totalProfit.dividedBy(maxDrawdown);
+        }
     }
 }
