+++ ta4j-core/src/main/java/org/ta4j/core/analysis/criteria/VersusBuyAndHoldCriterion.java
@@ -30,7 +30,7 @@ import org.ta4j.core.num.Num;
  * Versus "buy and hold" criterion.
  *
  * Compares the value of a provided {@link AnalysisCriterion criterion} with the
- * value of a {@link BuyAndHoldCriterion "buy and hold" criterion}.
+ * value of a {@link BuyAndHoldReturnCriterion "buy and hold" criterion}.
  */
 public class VersusBuyAndHoldCriterion extends AbstractAnalysisCriterion {
