+++ ta4j-core/src/main/java/org/ta4j/core/analysis/criteria/package-info.java
@@ -27,7 +27,7 @@
  * Analysis criteria can be used to compare two trading
  * {@link org.ta4j.core.Strategy strategy}.<br>
  * The most common criterion is the
- * {@link org.ta4j.core.analysis.criteria.TotalProfitCriterion total profit
+ * {@link org.ta4j.core.analysis.criteria.TotalReturnCriterion total return
  * criterion}. It measures how much is profitable a strategy.
  */
 package org.ta4j.core.analysis.criteria;
