+++ ta4j-core/src/test/java/org/ta4j/core/analysis/criteria/TotalReturnCriterionTest.java
@@ -1,0 +1,116 @@
+/**
+ * The MIT License (MIT)
+ *
+ * Copyright (c) 2014-2017 Marc de Verdelhan, 2017-2020 Ta4j Organization & respective
+ * authors (see AUTHORS)
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of
+ * this software and associated documentation files (the "Software"), to deal in
+ * the Software without restriction, including without limitation the rights to
+ * use, copy, modify, merge, publish, distribute, sublicense, and/or sell copies of
+ * the Software, and to permit persons to whom the Software is furnished to do so,
+ * subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all
+ * copies or substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+ * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS
+ * FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR
+ * COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER
+ * IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN
+ * CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+package org.ta4j.core.analysis.criteria;
+
+import org.junit.Test;
+import org.ta4j.core.AnalysisCriterion;
+import org.ta4j.core.BaseTradingRecord;
+import org.ta4j.core.Order;
+import org.ta4j.core.Trade;
+import org.ta4j.core.TradingRecord;
+import org.ta4j.core.mocks.MockBarSeries;
+import org.ta4j.core.num.Num;
+
+import java.util.function.Function;
+
+import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertTrue;
+import static org.ta4j.core.TestUtils.assertNumEquals;
+
+public class TotalReturnCriterionTest extends AbstractCriterionTest {
+
+    public TotalReturnCriterionTest(Function<Number, Num> numFunction) {
+        super((params) -> new TotalReturnCriterion(), numFunction);
+    }
+
+    @Test
+    public void calculateWithWinningLongTrades() {
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 105, 110, 100, 95, 105);
+        TradingRecord tradingRecord = new BaseTradingRecord(Order.buyAt(0, series), Order.sellAt(2, series),
+                Order.buyAt(3, series), Order.sellAt(5, series));
+
+        AnalysisCriterion profit = getCriterion();
+        assertNumEquals(1.10 * 1.05, profit.calculate(series, tradingRecord));
+    }
+
+    @Test
+    public void calculateWithLosingLongTrades() {
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 95, 100, 80, 85, 70);
+        TradingRecord tradingRecord = new BaseTradingRecord(Order.buyAt(0, series), Order.sellAt(1, series),
+                Order.buyAt(2, series), Order.sellAt(5, series));
+
+        AnalysisCriterion profit = getCriterion();
+        assertNumEquals(0.95 * 0.7, profit.calculate(series, tradingRecord));
+    }
+
+    @Test
+    public void calculateProfitWithWinningShortTrades() {
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 95, 100, 80, 85, 70);
+        TradingRecord tradingRecord = new BaseTradingRecord(Order.sellAt(0, series), Order.buyAt(1, series),
+                Order.sellAt(2, series), Order.buyAt(5, series));
+
+        AnalysisCriterion profit = getCriterion();
+        assertNumEquals((1 / 0.95) * (1 / 0.7), profit.calculate(series, tradingRecord));
+    }
+
+    @Test
+    public void calculateProfitWithLosingShortTrades() {
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 105, 100, 80, 85, 130);
+        TradingRecord tradingRecord = new BaseTradingRecord(Order.sellAt(0, series), Order.buyAt(1, series),
+                Order.sellAt(2, series), Order.buyAt(5, series));
+
+        AnalysisCriterion profit = getCriterion();
+        assertNumEquals((1 / 1.05) * (1 / 1.3), profit.calculate(series, tradingRecord));
+    }
+
+    @Test
+    public void calculateWithNoTradesShouldReturn1() {
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 95, 100, 80, 85, 70);
+
+        AnalysisCriterion profit = getCriterion();
+        assertNumEquals(1d, profit.calculate(series, new BaseTradingRecord()));
+    }
+
+    @Test
+    public void calculateWithOpenedTradeShouldReturn1() {
+        MockBarSeries series = new MockBarSeries(numFunction, 100, 95, 100, 80, 85, 70);
+        AnalysisCriterion profit = getCriterion();
+        Trade trade = new Trade();
+        assertNumEquals(1d, profit.calculate(series, trade));
+        trade.operate(0);
+        assertNumEquals(1d, profit.calculate(series, trade));
+    }
+
+    @Test
+    public void betterThan() {
+        AnalysisCriterion criterion = getCriterion();
+        assertTrue(criterion.betterThan(numOf(2.0), numOf(1.5)));
+        assertFalse(criterion.betterThan(numOf(1.5), numOf(2.0)));
+    }
+
+    @Test
+    public void testCalculateOneOpenTradeShouldReturnOne() {
+        openedTradeUtils.testCalculateOneOpenTradeShouldReturnExpectedValue(numFunction, getCriterion(), 1);
+    }
+}
