+++ ta4j-examples/src/main/java/ta4jexamples/analysis/StrategyAnalysis.java
@@ -52,27 +52,28 @@ public class StrategyAnalysis {
          */
 
         // Total profit
-        TotalProfitCriterion totalProfit = new TotalProfitCriterion();
+        TotalReturnCriterion totalProfit = new TotalReturnCriterion();
         System.out.println("Total profit: " + totalProfit.calculate(series, tradingRecord));
         // Number of bars
         System.out.println("Number of bars: " + new NumberOfBarsCriterion().calculate(series, tradingRecord));
         // Average profit (per bar)
-        System.out
-                .println("Average profit (per bar): " + new AverageProfitCriterion().calculate(series, tradingRecord));
+        System.out.println(
+                "Average profit (per bar): " + new AverageReturnPerBarCriterion().calculate(series, tradingRecord));
         // Number of trades
         System.out.println("Number of trades: " + new NumberOfTradesCriterion().calculate(series, tradingRecord));
         // Profitable trades ratio
         System.out.println(
-                "Profitable trades ratio: " + new AverageProfitableTradesCriterion().calculate(series, tradingRecord));
+                "Profitable trades ratio: " + new WinningTradesRatioCriterion().calculate(series, tradingRecord));
         // Maximum drawdown
         System.out.println("Maximum drawdown: " + new MaximumDrawdownCriterion().calculate(series, tradingRecord));
         // Reward-risk ratio
-        System.out.println("Reward-risk ratio: " + new RewardRiskRatioCriterion().calculate(series, tradingRecord));
+        System.out
+                .println("Reward-risk ratio: " + new ReturnOverMaxDrawdownCriterion().calculate(series, tradingRecord));
         // Total transaction cost
         System.out.println("Total transaction cost (from $1000): "
                 + new LinearTransactionCostCriterion(1000, 0.005).calculate(series, tradingRecord));
         // Buy-and-hold
-        System.out.println("Buy-and-hold: " + new BuyAndHoldCriterion().calculate(series, tradingRecord));
+        System.out.println("Buy-and-hold: " + new BuyAndHoldReturnCriterion().calculate(series, tradingRecord));
         // Total profit vs buy-and-hold
         System.out.println("Custom strategy profit vs buy-and-hold strategy profit: "
                 + new VersusBuyAndHoldCriterion(totalProfit).calculate(series, tradingRecord));
