+++ ta4j-examples/src/main/java/ta4jexamples/walkforward/WalkForward.java
@@ -28,7 +28,7 @@ import org.ta4j.core.BarSeries;
 import org.ta4j.core.BarSeriesManager;
 import org.ta4j.core.Strategy;
 import org.ta4j.core.TradingRecord;
-import org.ta4j.core.analysis.criteria.TotalProfitCriterion;
+import org.ta4j.core.analysis.criteria.TotalReturnCriterion;
 import org.ta4j.core.num.Num;
 import ta4jexamples.loaders.CsvTradesLoader;
 import ta4jexamples.strategies.CCICorrectionStrategy;
@@ -175,7 +175,7 @@ public class WalkForward {
         Map<Strategy, String> strategies = buildStrategiesMap(series);
 
         // The analysis criterion
-        AnalysisCriterion profitCriterion = new TotalProfitCriterion();
+        AnalysisCriterion profitCriterion = new TotalReturnCriterion();
 
         for (BarSeries slice : subseries) {
             // For each sub-series...
